Refactor test for make it easier to add new test
