+++ extensions/postgres/tests/postgres_test.go
@@ -1,66 +1,26 @@
-// Apache License 2.0
-
-// Copyright (c) XDBLab organization
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
+// Copyright 2023 XDBLab organization
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-//   http://www.apache.org/licenses/LICENSE-2.0
+// http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 package tests
 
 import (
-	"fmt"
-	"testing"
-	"time"
-
 	"github.com/stretchr/testify/assert"
-	"github.com/xdblab/xdb/common/log"
-	"github.com/xdblab/xdb/config"
-	"github.com/xdblab/xdb/extensions"
-	"github.com/xdblab/xdb/extensions/postgres"
-	"github.com/xdblab/xdb/extensions/postgres/postgrestool"
-
-	"github.com/xdblab/xdb/persistence/sql"
-	"github.com/xdblab/xdb/persistence/sql/basetest"
+	"github.com/xdblab/xdb/persistence/sql/sqltest"
+	"testing"
 )
 
-func TestPostgres(t *testing.T) {
-	testDBName := fmt.Sprintf("test%v", time.Now().UnixNano())
-	fmt.Println("using database name ", testDBName)
-
-	sqlConfig := &config.SQL{
-		ConnectAddr:     fmt.Sprintf("%v:%v", postgrestool.DefaultEndpoint, postgrestool.DefaultPort),
-		User:            postgrestool.DefaultUserName,
-		Password:        postgrestool.DefaultPassword,
-		DBExtensionName: postgres.ExtensionName,
-		DatabaseName:    testDBName,
-	}
+func TestBasic(t *testing.T) {
 	ass := assert.New(t)
 
-	err := extensions.CreateDatabase(*sqlConfig, testDBName)
-	ass.Nil(err)
-
-	err = extensions.SetupSchema(sqlConfig, "../../../"+postgrestool.DefaultSchemaFilePath)
-	ass.Nil(err)
-
-	store, err := sql.NewSQLProcessStore(*sqlConfig, log.NewDevelopmentLogger())
-	ass.Nil(err)
-
-	basetest.SQLBasicTest(ass, store)
-
-	_ = extensions.DropDatabase(*sqlConfig, testDBName)
-	fmt.Println("testing database deleted")
+	sqltest.SQLBasicTest(ass, store)
 }
