+++ persistence/sql/sqltest/base_basic.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package basetest
+package sqltest
 
 import (
 	"context"
